refactor: Project_analyzer Singleton 패턴 적용
